# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "barum 판정 백엔드",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "description": "헬스체크.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/reports/{result_id}": {
      "get": {
        "summary": "Get Report",
        "description": "저장된 검사를 다시 본다. 추측불가 result_id가 접근권이라 못 찾으면 404.",
        "operationId": "get_report_reports__result_id__get",
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredCheck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reports/{result_id}/image": {
      "get": {
        "summary": "Get Report Image",
        "description": "저장된 원본 이미지를 백엔드가 그대로 스트리밍(버킷 private, 서명 URL 없음).",
        "operationId": "get_report_image_reports__result_id__image_get",
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/check": {
      "post": {
        "summary": "Check",
        "description": "광고(이미지/글 + 나라)를 받아 문구별 위반 findings를 반환한다.\n\n이미지·글 중 최소 하나는 있어야 한다. 없으면 422. ingredients는 선택.",
        "operationId": "check_check_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_check_check_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/remediate": {
      "post": {
        "summary": "Remediate",
        "description": "위반 문구(sentence)와 유형(violation_type)을 입력받아 대체 표현을 제안한다.",
        "operationId": "remediate_remediate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemediationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemediationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_check_check_post": {
        "properties": {
          "region": {
            "$ref": "#/components/schemas/Region"
          },
          "ad_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ad Text"
          },
          "image": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image"
          },
          "ingredients": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingredients",
            "description": "전성분(콤마 구분). 있으면 2호 판정에 성분 정합 대조가 붙는다."
          }
        },
        "type": "object",
        "required": [
          "region"
        ],
        "title": "Body_check_check_post"
      },
      "CheckReport": {
        "properties": {
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings"
          },
          "unjudged": {
            "items": {
              "$ref": "#/components/schemas/UnjudgedSentence"
            },
            "type": "array",
            "title": "Unjudged"
          },
          "summary": {
            "$ref": "#/components/schemas/Summary"
          },
          "result_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Id"
          }
        },
        "type": "object",
        "required": [
          "findings",
          "summary"
        ],
        "title": "CheckReport",
        "description": "`POST /check` 응답. findings + unjudged + summary.\n\nresult_id: 이 검사가 저장됐으면 그 추측불가 id(다시 보기 URL). 저장 안 됐으면\n(JUDGE_KIND=stub·DB 미설정·저장 실패) None. 프론트는 있으면 다시 보기 링크를 건다."
      },
      "Finding": {
        "properties": {
          "span": {
            "type": "string",
            "title": "Span"
          },
          "sentence": {
            "type": "string",
            "title": "Sentence"
          },
          "violation_type": {
            "$ref": "#/components/schemas/ViolationType"
          },
          "legal_basis": {
            "type": "string",
            "title": "Legal Basis"
          },
          "flag": {
            "$ref": "#/components/schemas/JudgmentFlag"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          }
        },
        "type": "object",
        "required": [
          "span",
          "sentence",
          "violation_type",
          "legal_basis",
          "flag",
          "explanation",
          "location"
        ],
        "title": "Finding",
        "description": "문구 하나에 대한 판정 결과. 리포트에 한 줄로 렌더된다."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JudgmentFlag": {
        "type": "string",
        "enum": [
          "위반",
          "검토필요"
        ],
        "title": "JudgmentFlag",
        "description": "판정 확정도. v1.8: 위험도(고/중/저) 등급 폐지 → 위반/검토필요 이진화.\n\n'근거 있으면 위반, 근거 없으면 검토필요'(FR-5). RagJudge 이전엔 대부분\n유형에 규칙집 대조 수단이 없어 잠정적으로 위반으로 둔다(recall 우선).\n2호(기능성오인)만 예외 — 성분 정합이 실제 근거 대조라 그 결과를 쓴다."
      },
      "Location": {
        "properties": {
          "tile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tile"
          },
          "order": {
            "type": "integer",
            "title": "Order"
          },
          "y_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y Start"
          },
          "y_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y End"
          },
          "source_h": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source H"
          },
          "source_w": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source W"
          }
        },
        "type": "object",
        "required": [
          "order"
        ],
        "title": "Location",
        "description": "문구가 잡힌 위치.\n\nOCR은 글자 좌표(bbox)를 주지 않는다. 우리가 확보하는 위치 정보는 어느 타일의\n몇 번째 문장이냐까지다. 타일 단위 세로 밴드(y_start~y_end)를 원본 이미지\n좌표로 실어, 프론트가 원본 위에 밴드를 하이라이트할 수 있게 한다. 문장 단위\n정밀 좌표는 없다(OCR이 bbox를 안 주므로 타일 밴드가 최선).\n\n이미지 입력만 좌표가 있다. 텍스트 입력(이미지 없음)이면 tile과 좌표 모두 None.\nsource_h/source_w는 원본 이미지 크기라 프론트가 밴드를 원본 축척에 맞춘다."
      },
      "Region": {
        "type": "string",
        "enum": [
          "KR",
          "US"
        ],
        "title": "Region",
        "description": "검사 대상 국가. 지금은 국내(KR)만 실동작, 미국(US)은 2단계."
      },
      "RemediationRequest": {
        "properties": {
          "sentence": {
            "type": "string",
            "title": "Sentence",
            "description": "탐지된 위반 문구 전체"
          },
          "violation_type": {
            "$ref": "#/components/schemas/ViolationType",
            "description": "위반유형 라벨"
          },
          "span": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Span",
            "description": "위반으로 지목된 세부 표현 (생략 시 sentence 전체)"
          }
        },
        "type": "object",
        "required": [
          "sentence",
          "violation_type"
        ],
        "title": "RemediationRequest",
        "description": "수정 권고안 생성 요청."
      },
      "RemediationResponse": {
        "properties": {
          "sentence": {
            "type": "string",
            "title": "Sentence"
          },
          "violation_type": {
            "$ref": "#/components/schemas/ViolationType"
          },
          "span": {
            "type": "string",
            "title": "Span"
          },
          "suggestions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggestions",
            "description": "대체 표현 후보들"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "description": "권고 사항 고지 안내 문구"
          }
        },
        "type": "object",
        "required": [
          "sentence",
          "violation_type",
          "span",
          "suggestions",
          "disclaimer"
        ],
        "title": "RemediationResponse",
        "description": "수정 권고안 생성 응답."
      },
      "StoredCheck": {
        "properties": {
          "result_id": {
            "type": "string",
            "title": "Result Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "region": {
            "$ref": "#/components/schemas/Region"
          },
          "image_available": {
            "type": "boolean",
            "title": "Image Available"
          },
          "report": {
            "$ref": "#/components/schemas/CheckReport"
          }
        },
        "type": "object",
        "required": [
          "result_id",
          "created_at",
          "region",
          "image_available",
          "report"
        ],
        "title": "StoredCheck",
        "description": "`GET /reports/{result_id}` 응답 (다시 보기).\n\n저장된 CheckReport를 감싸고 저장 메타를 얹는다. 리포트 자체는 그대로라 프론트가\n\"방금 결과\"와 \"다시 보기\"를 같은 컴포넌트로 렌더할 수 있다. 원본 이미지는\n별도 프록시(`GET /reports/{id}/image`)로 받는다(image_available이 true일 때)."
      },
      "Summary": {
        "properties": {
          "region": {
            "$ref": "#/components/schemas/Region"
          },
          "n_sentences": {
            "type": "integer",
            "title": "N Sentences"
          },
          "n_findings": {
            "type": "integer",
            "title": "N Findings"
          },
          "n_violation": {
            "type": "integer",
            "title": "N Violation",
            "default": 0
          },
          "n_needs_review": {
            "type": "integer",
            "title": "N Needs Review",
            "default": 0
          },
          "n_unjudged": {
            "type": "integer",
            "title": "N Unjudged",
            "default": 0
          },
          "counts_by_type": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts By Type"
          }
        },
        "type": "object",
        "required": [
          "region",
          "n_sentences",
          "n_findings"
        ],
        "title": "Summary",
        "description": "리포트 상단 요약. 정책상 점수가 아니라 근거 개수로 표현한다.\n\nn_findings = n_violation + n_needs_review(합법·대상외 제외). n_unjudged는\n별개 개념(판정 자체를 못함, VLM 호출 실패) — 검토필요(판정은 했으나 근거\n약함)와 혼동하지 않는다."
      },
      "UnjudgedSentence": {
        "properties": {
          "sentence": {
            "type": "string",
            "title": "Sentence"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          }
        },
        "type": "object",
        "required": [
          "sentence",
          "location"
        ],
        "title": "UnjudgedSentence",
        "description": "판정하지 못한 문장.\n\nVLM 호출 실패(429·빈응답 등)로 위반 여부를 못 가린 문장. 정책상 recall 우선이라\n이걸 '합법'으로 삼키면 미탐이 숨는다. 그래서 findings에도 합법에도 넣지 않고\n별도로 남겨 '미판정(재검사 필요)'으로 드러낸다."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ViolationType": {
        "type": "string",
        "enum": [
          "합법",
          "1호_의약품오인",
          "2호_기능성오인",
          "5호_거짓과장기만",
          "대상외"
        ],
        "title": "ViolationType",
        "description": "화장품 광고 위반유형 (법 제13조 체계, 개정법 기준).\n\n화장품은 식품(1~5호)과 체계가 다르다. 3호는 삭제된 조항이라 없다.\n거짓·과장·기만은 현행법상 4호이나, 2026.11.27 시행 개정법에서 AI 생성물로\n전문가 보증을 오인시키는 광고 금지가 신설 4호로 들어오며 5호로 밀린다.\n시행 임박(발표 3주 뒤)이라 개정법 기준(5호)으로 미리 맞춘다. 신설 4호(AI)는\n문구 판정 라벨이 아니라 콘텐츠 생성 가드레일(FR-13) 영역이라 이 enum엔 없다.\n직렬화 값은 한국어 라벨(reference/cosmetic_kr 기준)."
      }
    }
  }
}